Render/defect-detection-app/requirements.txt
streamlit==1.45.1
tensorflow==2.19.0
numpy
pillow